# order-service/openapi.yaml
openapi: 3.0.0
info:
  title: Order Service API
  description: Order Management and Tracking Microservice
  version: 1.0.0
  contact:
    name: CTSE Team
    url: https://ctse.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3003
    description: Development server
  - url: http://order-service:3003
    description: Docker network (internal)
...
paths:
...
  /api/orders:
    get:
      summary: Get all orders with pagination and filtering
      tags:
        - Order Management
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        - name: userId
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  pages:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      summary: Create a new order
      tags:
        - Order Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - items
                - shippingAddress
              properties:
                userId:
                  type: string
                  example: "550e8400-e29b-41d4-a716-446655440000"
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                        minimum: 1
                shippingAddress:
                  type: object
                  properties:
                    fullName:
                      type: string
                    email:
                      type: string
                    phone:
                      type: string
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    postalCode:
                      type: string
                    country:
                      type: string
                billingAddress:
                  type: object
                paymentMethod:
                  type: string
                  enum: [credit_card, debit_card, paypal, bank_transfer, other]
                notes:
                  type: string
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or insufficient inventory
        '404':
          description: User or product not found
...
components:
  schemas:
    Order:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        userId:
          type: string
        userEmail:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              sku:
                type: string
              productName:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
              totalPrice:
                type: number
        summary:
          type: object
          properties:
            subtotal:
              type: number
            tax:
              type: number
            shipping:
              type: number
            discount:
              type: number
            total:
              type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        statusHistory:
          type: array
        tracking:
          type: object
          properties:
            currentStatus:
              type: string
            trackingNumber:
              type: string
            carrier:
              type: string
            events:
              type: array
        shippingAddress:
          type: object
        billingAddress:
          type: object
        payment:
          type: object
          properties:
            method:
              type: string
            status:
              type: string
            transactionId:
              type: string
            paidAt:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time